+++ uni_modules/echarts-for-uniapp/components/uni-chart/uni-chart.vue
@@ -14,7 +14,18 @@
 
   <!-- #ifdef MP-WEIXIN || MP-QQ -->
   <canvas
+    v-if="useNewCanvas"
     type="2d"
+    :id="canvasId"
+    class="uni-canvas"
+    :canvas-id="canvasId"
+    @click="emits('click', $event)"
+    @touchstart="touchStart"
+    @touchmove.stop="touchMove"
+    @touchend="touchEnd"
+  />
+  <canvas
+    v-else
     :id="canvasId"
     class="uni-canvas"
     :canvas-id="canvasId"
@@ -42,6 +53,14 @@ let chart
 export default {
   name: 'uni-chart',
   props: {
+    forceUseOldCanvas: {
+      type: Boolean,
+      default: false
+    },
+    theme: {
+      type: String,
+      default: ''
+    },
     option: {
       type: Object,
       require: true,
@@ -54,20 +73,18 @@ export default {
     return {
       canvasId: `uni-canvas-${Date.now()}`,
       systemInfo: uni.getSystemInfoSync(),
+      useNewCanvas: true
     }
   },
-  mounted() {
-    // console.log(uni.getSystemInfoSync())
-    this.$nextTick(() => {
-      this.registerPreprocessor()
-      // #ifdef APP-VUE || H5
-      this.renderH5()
-      // #endif
+  onReady() {
+    this.registerPreprocessor()
+    // #ifdef APP-VUE || H5
+    this.initH5()
+    // #endif
 
-      // #ifdef MP-WEIXIN || MP-QQ
-      this.renderMpWeixin()
-      // #endif
-    })
+    // #ifdef MP-WEIXIN || MP-QQ
+    this.initMiniProgram()
+    // #endif
   },
   methods: {
     // register echarts preprocessor
@@ -84,14 +101,44 @@ export default {
         }
       })
     },
-    // renderH5
-    renderH5() {
+    // init H5 app-vue
+    initH5() {
       const canvasNode = document.getElementById(this.canvasId)
-      chart = echarts.init(canvasNode)
+      chart = echarts.init(canvasNode, this.$props.theme)
       chart.setOption(this.$props.option)
     },
-    // render mp-weixin
-    renderMpWeixin() {
+    // init mini program
+    initMiniProgram() {
+      const version = this.systemInfo.SDKVersion
+      console.log(`当前基础库版本为: ${version}`)
+
+      const oldVersion = '1.9.91'
+      const baseVersion = '2.9.0'
+
+      let canUseNewCanvas = this.compareVersion(version, baseVersion) >= 0
+
+      if (this.$props.forceUseOldCanvas) {
+        if (canUseNewCanvas) console.warn('开发者强制使用旧canvas,建议关闭')
+        canUseNewCanvas = false
+      }
+      this.useNewCanvas = canUseNewCanvas && !this.forceUseOldCanvas
+      if (this.useNewCanvas) {
+        // 2.9.0 可以使用 <canvas type="2d"></canvas>
+        console.log(`基础库版本大于${baseVersion}，开始使用<canvas type="2d"/>`)
+        this.initNewCanvas()
+      } else {
+        const isValid = this.compareVersion(version, oldVersion) >= 0
+        if (!isValid) {
+          console.error(`基础库版本过低，需大于等于 ${oldVersion}。`)
+          return
+        } else {
+          console.warn(`建议将基础库调整大于等于${baseVersion}版本。升级后绘图将有更好性能`)
+          this.initOldCanvas()
+        }
+      }
+    },
+    // initNewCanvas
+    initNewCanvas() {
       const query = uni.createSelectorQuery().in(this)
       query
         .select(`#${this.canvasId}`)
@@ -103,7 +150,7 @@ export default {
           ctx.scale(this.pixelRatio, this.pixelRatio)
           const canvas = new UniCanvas(ctx, canvasNode)
           echarts.setPlatformAPI({ createCanvas: () => canvas })
-          chart = echarts.init(canvas, '', {
+          chart = echarts.init(canvas, this.$props.theme, {
             width: canvasNode.width,
             height: canvasNode.width,
             devicePixelRatio: this.pixelRatio
@@ -112,6 +159,32 @@ export default {
         })
         .exec()
     },
+    // initOldCanvas
+    initOldCanvas() {
+      // 1.9.91 <= sdkVersion < 2.9.0：原来的方式初始化
+      const ctx = uni.createCanvasContext(`#${this.canvasId}`, this)
+      const canvas = new UniCanvas(ctx)
+      echarts.setPlatformAPI({ createCanvas: () => canvas })
+      const query = uni.createSelectorQuery().in(this)
+      query
+        .select(`#${this.canvasId}`)
+        .boundingClientRect(res => {
+          // 微信旧的canvas不能传入dpr
+          const pixelRatio = 1
+          const canvasWidth = res.width
+          const canvasHeight = res.height
+
+          console.log(canvasWidth, canvasHeight)
+          chart = echarts.init(canvas, this.$props.theme, {
+            width: canvasWidth,
+            height: canvasHeight,
+            devicePixelRatio: pixelRatio
+          })
+          chart.setOption(this.$props.option)
+        })
+        .exec()
+    },
+    // init
     // event
     wrapTouch(event) {
       for (let i = 0; i < event.touches.length; ++i) {