+++ uni_modules/echarts-for-uniapp/changelog.md
@@ -1,2 +1,17 @@
+## 1.0.1（2022-11-19）
+- 新增 支持vue2.x, 旧版微信小程序
+- 新增 h5点击事件
+- 新增 小程序触摸事件
+- 新增 forceUseOldCanvas参数：强制使用旧版Canvas
+- 移除 width、height属性，用style代替
 ## 1.0.0（2022-11-18）
-在uni-app中可以以组件的形式使用 ECharts, 支持touch事件。
+
+- 在uni-app中可以以组件的形式使用 ECharts, 支持touch事件。
+
+## 1.0.1 (2022-11-19)
+
+- 新增 支持vue2.x, 旧版微信小程序
+- 新增 h5点击事件
+- 新增 小程序触摸事件
+- 新增 forceUseOldCanvas参数：强制使用旧版Canvas
+- 移除 width、height属性，用style代替